Activate admission webhook
Add the DeclaredVersionHydrator which populates the declared API
Version for resources.

Make the admission webhook actually run by removing the "allow all"
logic.

Fix issue where we didn't properly wrap errors related to YAML parsing.

Fix test which previously implicitly declared an empty set of rules for
an aggregate ClusterRole.

Only create admission webhook deployment for multi-repo. There's a few
changes we'd have to do to monorepo to make it support multi-repo, so
it's better to just ignore it.

The rest of the CL is just test changes.

Bug: 162496882
Bug: 179778079
Bug: 182467861
Bug: 182468449

Change-Id: I3066b35bf53a27a27c1a6fe1dbdda067fcae78aa