Delete webhook on uninstall
Rather than dynamically creating the webhook, declare it in the
manifest. This empty entry has no effect until populated with rules, so
it doesn't impact anything by default.

By having the declaration in the CSMR manifest, the Operator will
automatically delete it on uninstall.

Also don't add metadata.namespace to the webhook since webhook
configurations don't have Namespace.

The alternative would be to tell the Operator to explicitly remove this
object specifically, which is brittle.

Bug: 182593817
Change-Id: I3a881b8b3b17f94a4164414a319b2b8d97bcbefb