Set default port protocol
Due to a known issue in k8s, getting the set of declared fields for
anything with a []ContainerPort or []ServicePort results in an error.
This CL automatically sets the protocol to "TCP" in the case that the
user didn't set it explicitly.

The main point of the logic here is making "encodeDeclaredFields" set
these fields when they're unset. Most of the types hold ports in the
same place, but there are a few outliers which are handled.

We ignore ephemeralContainers even though they contain the
ports field as:
1) it makes no sense to declare ephemeralContainers in git, and
2) it is forbidden to declare ports for ephemeralContainers.

A lot of this logic involves messy unstructured.Unstructured logic -
this is unavoidable. Users shouldn't run into these errors as they'd
appear earlier during parsing. If in practice these errors are common
we'll have to rethink it.

Add exact YAML tests for these cases. In all cases I've ensured that
without the change the tests fail and with the change they pass. Also
add an E2E test to ensure we know if we accidentally mess this up later.

Bug: 182593448

Change-Id: I8a10ec551d985d91a354771f4881fee42f065cb1