Ignore the PruneType event error when the type is not found
When the CR is removed from the repo after the CRD is deleted,
the CR has been deleted from the cluster when the CRD is deleted.
The reconciler shouldn't report any error. Due to the bug in
kpt live apply library, there is an PruneType event emitted
for that CR, which cause an error being present in RepoSync status.

This change is to ignore that error from the PruneType event
when the resource type is not found.

Bug: b/183150103
Change-Id: I3d27af9ddd0f5f143cb94fcfddc02a8c7475fcaa