Modify logic in otel controller to prevent operator from overwriting changes to otel-collector config map.
Previously, the otel controller would patch the standard `otel-collector` ConfigMap with a Stackdriver config if Application Default Credentials were found. However, because this CM was controlled by the Operator, it would revert that patch and lead to a resource fight, causing an endless loop of updates.

Rather than patch the standard CM, the otel controller now creates a new Stackdriver ConfigMap.

Bug: 183215999
Change-Id: Ic70f37f3d9362732e5f21f10d8cd7252fd588455