Make FileObject use unstructured
The big change here is moving from client.Object in FileObjects to
unstructured.Unstructured. This has a few minor implications for how we
reference and copy the underlying object, but it's mostly changing
".Object" to ".Unstructured". This rids of us of several unnecessary
type conversions.

Unfortunately it requires a few painful changes to tests. The type
system is very inconsistent when deciding whether to include empty
annotations and labels maps, so I've force-included these whenever we
create an Unstructured for tests, and when reading from disk. This
slightly alters the "declared fields" annotation, but the difference is
only cosmetic. Because of the structs-cant-be-nil-to-unstructured
problem, I've had to change several tests from Role/RoleBinding to
ConfigMap since ConfigMamps don't have the same problem.

Bug: 181994737

Change-Id: I234c0b1d69d0e86dc6b2c5140e3e1662f60b7964