Delete admission-webhook in the clean step
The existence of the admission-webhook prevents deleting the test
resources, which results in a polluted testing environment. When running
the CI tests in a shared GKE cluster, tests fail because of test
leftovers.

This commit deletes the webhook before and after each test to make sure
the shared testing environment is clean.

b/183612179

Change-Id: Iceda6eb44b0fa52a2ac30821a40938fc88187ba2