Use a longer REST config timeout for the remediator
The current REST config timeout (5 seconds) causes all the watches to
restart every 5 seconds. This can be expensive because the newly started
watch needs to relist all the resources.

Get a separate config for the remediator to talk to the apiserver since
we want a longer REST config timeout for the remediator to avoid
restarting watches too frequently.

b/184064487

Change-Id: I4aad3cc095c596933e48c9b04bc6fbb60cf2d4aa