Fix the e2e test failures caused by restarting otel-collector
There're two issues causing the otel-collector being restarted.
1. in the production code
When reconcielr-manager restarts, the otel-collector deployment
in the config-management-monitoring namespace is also restarted
because the hash of otel-collector-stackdriver configmap is changed.
However, the content of the configmap is not updated. The hash is
changed because the orders of all fields are different. To fix that,
if CreatOrUpdate(configmap, ...) returns `unchanged`, we don't
return the hash, but a nil, which will not update the otel-colelctor
deployment.
2. in the e2e test code
The otel controller creates or updates the `otel-collector-stackdriver`
configmap, which causes the otel-collector deployment to be updated.
The e2e tests are intermittently hitting the port-forwarding issue as
below:
```
nothing written to stdout for kubectl port-forward, stdout="",
stderr="error: error upgrading connection: unable to upgrade connection:
pod not found
(\"otel-collector-57f8fd9bc7-vw8ww_config-management-monitoring\"
```
This commit guarantees that the e2e tests are started after the
otel-collector deployment has already been updated, so that we won't hit
the error anymore.

The above are known reasons that otel-collector gets restarted, but if
it is restarted during the long-running test for other unknown reasons,
we should be able to handle the port update.
This commit adds retry to PortForwardOtelCollector to make sure the port
is updated if otel-collector restarts.

It also adds `apply_operations_total` metric to the validatedMetrics
list so that new metrics can be appended.

b/183962428

Change-Id: If48042634e2912c47f34ea7f35336c4fd4d2dcb1