Decrement the watches metric when runWatcher returns errors
We increment the watches metric when starting a new watcher and
decrement when stopping a watcher. When runWatcher returns errors, we
delete the gvk from the watcherMap similarly as stopWatcher. We should
decrement the watches metric in this case as well.

The current presubmit check didn't catch this issue because each test
case has a brand new otel-collector, which is hard to accumulate the
watches metrics. The fix was verified in the failing CI tests when
using a shared GKE cluster.

b/183962428

Change-Id: I6d14089ab817d04a9388f1f973090b6585c7562d