Fix --no-api-server-check in nomos hydrate
tg/979833 was a large refactor of how we use Scoper to validate
resources based on the API Server state. Specifically, it moved the
usage of the flag for skipping the API Server out of the parser and into
the main logic of hydrate.go.

In this transition, we overlooked that this turned the conditional call
to get CRDs from the cluster into one that happens no matter how
--no-api-server-check is set.

This CL fixes the issue by making GetSyncedCRDs explicitly ask for the
flag, so it cannot be forgotten. It isn't strictly the simplest
solution, but I belive it will reduce the likelihood of such an error in
the future.

Also adds an E2E test that ensures --no-api-server-check works. I've
tested this CL both manually and in the E2E test and the flag works as
expected now.

Bug: 184803213
Change-Id: I66a2b8e1ed608ebdb8f0948f9b4dc413d57bd317