Prevent DiscoveryClient creation when --no-api-server-check
Not every function respects --no-api-server-check, and it can be
arduous to pass this flag through the entire call stack. This caused
problems in certain instances of CRs in repositories with no
corresponding CRD where we would talk to the API Server anyway.

This CL circumvents the issue by simply never creating a DiscoveryClient
if --no-api-server-check is passed, we don't call any methods which
require a full DiscoveryClient, and for ServerResourcer we pass a new
NoOpServerResourcer which does nothing.

Bug: 184803213
Change-Id: I5957a45afa6b9749c3c0579bcfbfe71376975caa