Make nomos hydrate output pass vet
As-is we add annotations/labels to objects that make the output of nomos
hydrate unsuitable for using with Config Sync.

This CL makes nomos hydrate automatically remove unsuitable annotations
and labels from objects before writing to the disk. It does this by
exposing the validation logic for validating labels/annotations, and
removing illegal ones. Note that if logic is added outside these
methods, it may be possible for this issue to recur.

Also, add CLI test to ensure this continues to work.

Remove the annotations:
- configmanagement.gke.io/cluster-name
- configmanagement.gke.io/source-path

Remoe the label:
- configsync.gke.io/declared-version

Bug: 188901354
Change-Id: I15dba2527140a792e1ee1b68d4a12bd4f4481e0e