Fix the url format issue in nomos status output
The nomos status output shows the repo url with `https:/...` because it
uses path.Join() to join the repo url and policy dir. `path.Join()`
invokes a Clean() function which replaces multiple slashes with a single
slash.

This commit replaces path.Join() with string concatenation. It also
trims the tailing slash in the repo url and the leading slash in the
policy dir.

b/190538584

Change-Id: Idd36d17509919ad109a64229d5835abd9392402c