Avoid pushing the `configmanagement.gke.io/managed: disabled` annotation to the cluster
By design, the `configmanagement.gke.io/managed: disabled` annotation
should not be pushed to the cluster. Instead, we remove all the Config
Sync metadata from the object on the cluster.

Pushing this annotation to the cluster adds extra work to our
remediator, which tries to remove this annotation since the management
for this object is disabled.

Change-Id: I4dffb88e858a1d969d941dc74c9ab1585ceb1559