Compute the reconciler deployment checksum after mutation
Previously, we compute the reconicler deployment hash before the
mutation, and it was always set to the checksum of the original
template. It causes issues when we switch from
`none/gcenode/gcpserviceaccount` to other secret types because the
deployment with the former types has the same checksum and the
deployment is not patched.

This commit computes the checksum after mutation. Deployment spec
updates caused by different secret types will be reflected in the
checksum. Hence, the deployment will be updated.

b/193032001

Change-Id: I647819a19c1866974ca85a97f85dbbc58d39a94b