Discover git-sync failures after successful syncs
If git-sync fails to sync from a repo that was previously synced
successfully, git-sync would report an error, but the reconciler
container won't be able to capture the error because the shared volume
still has the previous copy. This causes the force-resync to fail:
b/193111140.

This commit leverages the git-sync error file to discover the sync
failure.

b/193111140

Change-Id: I1148a71d394fbecced92236100ef771aea4f6360