Add rendering status to the `.status` field
It is likely Config Sync is rendering a different commit than the
imported source commit. Originally, we thought about adding the
rendering status as conditions, but we decided to not do so for the
following reason:
- it's difficult to distinguish between a successful rendering and a
rendering that has not been started yet.
- it's error prone to fetch the commit hash from the message in the
condition.

Now the rendering status will be a subfield under `.status`, which
includes the commit hash of the source configs that are rendered, the
rendering phase (skipped, in-progress, succeeded, or failed), and
rendering errors.

b/194301093

Change-Id: If8d980b5a3cde95c63d781ed58906f7502219da2