Fix the failure in the CI job for psp cluster
The namespace reconciler needs a clusterrolebinding to run
on a PSP cluster. When setting the ClusterRoleBinding file path
in the cnetralized mode, it was saved in a file crb.yaml.
In the testcases where two namespaces are used(backend, frontend),
this file will be overwritten. So the ClusterRoleBinding for
the prior namespace is missing.

Change-Id: I15a5c0abab9b9b286974aa25442c832435be04c2