Skip running bats tests on non-testing git providers
The bats tests use a hard-coded repo URL, which connects to a test
git-server. It requires a lot of add-hoc changes in the bash scripts
to set the URL dynamically for different git providers with random
repo names. Most of the bats e2e tests have been converted into
golang tests, so it should be safe to skip the bats tests when
running e2e on non-testing git providers.

b/194139809

Change-Id: Iac4146dee0bb80e2291b465072f070c7c54ebf5b