Do not check if the source directory is clean or not in multi-repo mode
Currently the reconciler returns a source error if the source git
directory is not clean. This is to prevent Config Sync from syncing
configs that are not checked in the git repository.

However, the hydration controller might pull in the remote Helm charts
to the local git directory, which will make the check fail. There are
two possible workarounds:
1. check if the git directory is clean before rendering, and reset the
git directory after rendering is done. This approach guranteens that the
git directory is clean when the next rendering is kicked in. But because
the local charts, we will see transient errors in the rendering process.
2. keep the source git directory read-only, copy the content to a temp
directory, and render the temp directory. This will avoid the transient
error because the source directory is not changed, but it brings latency
if the source directory is big. Moreover, the temp directory can still
be modified before Config Sync parses the configs.

Given neither of the above approaches are perfect, we decided to remove
the check. Ideally, the customers should have limited access on the
volume, which prevents them from modifying the config files.

b/194301093

Change-Id: I79d96b414abef858e6b498d86ba584aeaee422c2