Add .gitStatus field to rendering status and sync status
Currently, only `.status.source` field includes the Git info in the
`.gitStatus` subfield. It is causing confusion if the repo being
rendered is different from the one being parsed, or being synced.

This commit adds the `.gitStatus` field to the rendering status and sync
status to avoid such confusion.

b/196413339

Change-Id: I33bad24ff2034f4a7571d84ae4f27115621afa1d