Add the lastUpdate field to the rendering and source statuses
With the .status.rendering field added to the RootSync and RepoSync API,
it is likely that the rendering commit, source commit and sync commit
are all different. It is not obvious to know which commit is the latest.
Hence, we decided to add the lastUpdate field to each of the sub field
to indicate the execution order.

b/196437244

Change-Id: I250b37c4f34a873b80c8f989370eb58c2e46996e