Update 'nomos status' to reflect the rendering status
With the new status.rendering field added to the API, the calculation
needs to be updated to incorporate the rendering status, source status
and sync status.
The new Config Sync status will be calculated as below:
- SYNCED: .status.rendering.commit is equal to .status.sync.commit, and
no rendering errors, source errors, or sync errors.
- FAILED: rendering errors, source errors, or sync errors are found.
- PENDING: .status.rendering.commit is not equal to .status.sync.commit,
and no errors (rendering in progress, source parsing in progress, or
syncing in progress).

More details:
http://docs/document/d/1P7c58c6u22rv3vX6ZgiJl4b3G_rDgjN-48vq4GVW7Kg?resourcekey=0-tyEieTqLiNxDfcnmfR30NQ#heading=h.ilm0lpzbm4rg

b/196832460

Change-Id: Ic18d0653e21bfc6fb07af5293a32f072e4107215