Add commit hash to the hydration done file to avoid repeated hydration
The hydration controller extracts the hydrated commit hash from the hydrated
directory and compares it with the source commit. If they are the same,
it will not hydrate the samme commmit. However, the hydrated commit hash
is only extracted from a successful execution. If the hydration fails,
it will keep hydrating and keep failing in each hydration cycle.
This will cause the rendering status to bounce betwen in-progress state
and failed state.

To make hydration failed status consistent, this CL adds the commit hash
to the done file. In the beginning of each hydration, it checks if the
hash in the done file is the same as source commmit hash. If they're the
same, regardless of success of failure, it will skip processing.

The commit also adds a rehydrate option to retry on errors. If the
hydration failure is transient, we definitely don't want to prevent
hydation-controller from recovering. The rehydrate option will rerun the
hydration for failure causes every hour.

This also moves the update of the SkipRenderingCount and RenderingCount
metrics to the reconciler when updating the rendering status. It
guarantees the metrics are only updated once for each commmit.

Change-Id: I2ec0063c1cb7e5bf4531fdafa0b1c65e53bf1fac