Detect kustomization.yaml in subdirectories
If a DRY repo doesn't have a kustomization.yaml file in the root
directory, Config Sync returns a list of source errors for unknown kind
of kustomization.yaml files.

This commit detects kustomization.yaml files in subdirectories. If
found, but no kustomization.yaml file in the root directory, it returns
a user actionable hydration error with error code KNV1068.

b/198176650

Change-Id: Id3568411bed4354603e34ea9eb1b71bad4132888