Fix the missing config issues for `nomos hydrate`
There are two issues of `nomos hydrate`:
1. configs for clusters selected by the cluster-name-selector annotation
are not generated.
Currently, `nomos hydrate` hydrates default configuration for
`defaultcluster`, and hydrate specific configuration for clusters
selected by the ClusterSelector objects. However, the clusters selected
by the `cluster-name-selector` annotation are not hydrated.
This commit filters the clusters selected by both the ClusterSelector
objects and `cluster-name-selector` annotation.

2. `nomos hydrate` reads the Clusters object from the `clusterregistry`
directory to generate directories for selected clusters. However, the Cluster
objects can exist in any directory for unstructured format.
Hence, this commit updates the parser to read all file paths for
unstructured format.

This commit adds a bunch of test examples:
1. hierarchical repo with the cluster-name-selector annotations. Cluster
objects don't exist in this example, which verifies configs can be
generated for clusters selected by the cluster-name-selector annotation.
2. unstructured repo with the ClusterSelector, Cluster objects. This
verifies the Cluster objects can be put to any arbitrary location.
3. unstructured repo with the cluster-name-selector annotations, but no
Cluster objects. Similarly as the hierarchical repo, this is to verify
configs can be generated for clusters selected by the
cluster-name-selector annotation regardless of the source format.

b/200068373
b/200214126

Change-Id: I8916300531d22ba8c936a3ef45f2ed1990e80071