Avoid change the status of a namespace from active to inactive
Background
Config Sync decides the activeness of a namespace based on whether it
has the its ClusterSelector annotation or the value of its
ClusterSelector annotation: a namespace is active if it does not have
the ClusterSelector annotation, or its ClusterSelector annotation
includes the cluster name;

Issue
If a namespace is defined multiple times with different ClusterSelector
annotation, currently its activeness is determined by the last
definition. For example, if the current cluster name is `cluster1`, and
the Git repo defines a namespace twice: the ClusterSelector annotation
for the first definition is `cluster1`, and ClusterSelector annotation for
the second definition is `cluster2`. After processing the first
definition, we mark the namespace as active; however, after processing
the second definition, we mark the namespace as inactive.

Impact of the issue
This issue causes Config Sync only applies the resources under the
affected namespace to one cluster, even if these resources are declared
for multiple clusters using ClusterSelector annotations.

Fix
The fix is to avoid change the status of a namespace from active to
inactive.

b/200182865
b/199511373

Change-Id: I365b3d64304cdf5be036d6b2559f7130a0abee21