Increase git-importer default memory limit to 500Mi
Given the OOM errors encountered with Gatekeeper, a 200Mi limit is
simply not enough.  See b/201146483 and cl/399207287.

Ideally we want to set requests and limits closer to each other, and
ideally we want to reduce the memory footprint (it seems to be related
to watching Constraints and ConstraintTemplates).  However, as a first
step, we we can simply increase the default memory limit, meaning
fewer customers should encounter OOM errors.

b/201146483

Change-Id: I4bc73679a1ed374254c54a683868d838484a0581