Make RenderingInProgress not an error
When the rendering status is not ready, the reconciler logs an error and
clears the state tracking information. There are customers monitoring
the errors in the reconciler log. The RenderingInProgress is misleading,
and should not be an error-level log. This commit updates to make it an
info-level log.

It also makes the reconciler to retry every second instead of
exponential backoff when the rendering status is not ready.

b/202015971

Change-Id: I260efc3583375a77fbf8568f2a214421266d0e49