Avoid hard-coding the k8sallowedrepos.constraints.gatekeeper.sh CRD in Config Sync
http://tg/574337 added support for handling Constraint and
ConstraintTemplate in same commit, which hard-coded the
k8sallowedrepos.constraints.gatekeeper.sh CRD even if it is neither
installed on the cluster or declared in the Git repo.
This would cause an k8sallowedrepos.constraints.gatekeeper.sh CR to be
sent to the kpt applier before the CRD is installed on the cluster or
included in the Git repo: https://team.git.corp.google.com/gke-kubernetes-enterprise-control/prototype/+/refs/heads/master/e2e/testcases/gatekeeper.bats#36.

This works well with the cli-utils version used in Config Sync 1.9.1,
however will not work with the new cli-utils version, which would
terminate if unknown-scoped objects are encountered.

This is to prepare for updating the cli-utils library in Config Sync.

b/202180531

Change-Id: Id93b1df4e47b79042a8bbf5a249fce88f73e8151