Add pipeline error metric to reflect error syncing commit
This is part of [observability][1] that adds metric from
Config Sync side to reflect error when syncing a commit.
The change consolidates the hydration errors as one of the
components.

This change also changes metric tagkey KeyReconciler to use
a fixed string as [this issue][2] was resolved, so that thi
label can be correctly defined in Cloud Monarch.

[1]: http://shortn/_xY9qiO2vOh
[2]: http://shortn/_Ueyn1L0ZxY

b/200074182 b/199790294

Change-Id: I13f712bbc682b4d719b04028ee6c7b36a0c41da4