Update `Status.Source` immediately after the configs are parsed
Currently, after the configs are parsed, we don't update `Status.Source`
immediately. This is to avoid updating the RootSync/RepoSync object
twice in a short time if the apply step finishes fast.

However, if the apply step takes a long time to finish, `Status.Source`
may be empty for quite a while (if this is the first reconciliation
after a reconciler starts) or stale (this causes `Status.Source.Commit`
to be older than `Status.Rendering.Commit` for a long time, which can be
confusing and makes people think the configs failed to be parsed).

Change-Id: Icffbabd3527e75229c966a02b364ecc2aae44997