Avoid copying source configs to a temp folder in 'nomos hydrate'
Currently, nomos hydrate copies the source configs in the --path
directory to a temp folder in case kustomize build generates additional
files, which pollutes the source files. However, the --path directory may
not include all the resources needed by kustomize.

We should avoid copying files and document the files that might be
generated by kustomize. Instead, we generate a warning in the command to
remind users that a 'charts' folder might be generated.

b/202976017

Change-Id: I8129b57dea43ef9b14225b93aa18a1348b9146b2