Increase memory request for git-sync container.
This is necessary so that GKE autopilot does not auto-clamp the memory
limit down so low that the container can not function. Looking at
current memory utilization indicates that 99.9% of fleet is using 170Mi
or less, so requesting 200Mi makes sense (and that matches the limit).

For further work, we could update all of the memory and CPU
requests/limits to match each other for all containers since that will
happen automatically with GKE autopilot. This fix should unblock the
critical issue to continue testing autopilot compatibility.

Bug: 202906842
Change-Id: I036f96a09a4d65fa122bf4e0685dd33867aeb0da