Make reconciler-manager revert changes to the reconciler Deployment
http://tg/1069136 introduces the Deployment manifests checksum
annotation and uses the annotation to update the Deployment manifests.
If the checksum of the current reconciler Deployment is the same as the
one in the ConfigMap, it copies the current Deployment and patches based
on that. One issue is that the Deployment can be manually edited while
the checksum remains the same. This results in the issue that the
reconciler manager doesn't revert changes to the reconciler Deployment.

To revert the changes, we can't rely on the checksum. Instead, we need
to compare the spec and labels directly.

To support vertical autoscale by autopilot, changes to the container
resource requirements need to be ignored when the autopilot annotation
exists.

b/203700215

Change-Id: I2de1a74a3679788d0c1809b7c3de5a599d6204ed