Update open telemetry image
Config Sync was using a custom otel image with third-party license
files. We need to update to the latest version of open telemtry with
third-party licenses.

The updated image is also a custom build on top of
opentelemtry-collector-contrib:v0.38.0 with licenses added.

Both the agent and the collector use the same image because the
opencensus receiver and exporter are removed from
opentelemetry-collector:
https://github.com/open-telemetry/opentelemetry-collector/issues/3474.
The healthcheck extension is also removed from the
opentelemetry-collector:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/4894.

b/195486045

Change-Id: Ie7649e0b937a1e4c775ff557f33614c4159c596a