Set the default source-format to "" for nomos commands
`nomos vet` use different default source formats. In most cases, if
source-format is not set, it defaults to `hierarchy` for backward
compatibility. But when the `--namespace` flag is set, it defaults to
`unstructured` to be able to validate a namespace repo.

Hence, this commit updates the default value from "hierarchy" to "" to
allow different defaults.

b/206506531

Change-Id: I2cc14d30cc7ea96b875cff93558667f5f6392edf