Increase memory request for reconciler container
The e2e tests fail constantly on the Autopilot clusters because
the reconciler gets OOMKilled with a low memor limit controlled by GKE
Autopilot.

The current utilization shows that 95% of fleet is using around 200Mi
memory: http://screen/7cibWgrkskG5YZb. So, this CL increases the memory
request of the reconciler container to 200Mi.

b/205869701

Change-Id: I26f3a26f66067e8010378398556c787437316fe4