Skip e2e tests that mutate kube-system resources on Autopilot clusters
kube-system is a managed namespace in the GKE Autopilot clusters. If
Config Sync syncs a kube-system resource, it will fail with a
GKEAutopilot authz error.

This CL detects if the tests are running on an Autopilot cluster and
skip the testcases that mutate the kube-system resources.

It also increases the default wait duration from 2 minutes to 6 minutes
when running on Autopilot clusters because Autopilot needs more time to
spin up Pod.

b/205869817

Change-Id: Ife2f27e1a3cbfdbbd0d550948692a426d4717799