Update the sync status periodically during apply
Currently, a Config Sync reconciler only updates the sync status after
the apply step is done. If the apply step takes a while, the
`status.sync` field is empty during apply, and the `message` field of the
`Syncing` Condition is `Source` during apply.

This PR updates the sync status every 5 seconds and sets the `message`
field of the `Syncing` Condition to `Syncing` during apply.

b/205874856

Change-Id: I26cccbc66fffa25d1b36470ee6b57432c65de9c0