Update the helm examples
The original examples include mutating webhook objects in the helm
charts, which fail to be applied on Autopilot clusters.

This CL updates the helm examples to use new charts that
don't have mutating webhook objects.

It also switches the Kustomization configuration to use the config of
the default helm plugin as that's the documented behavior. The
render-helm-chart function is used as an alternative when the default
helm plugin is deprecated. However, it is current hitting an issue on
the autopilot cluster: b/209458334. So tests around the function are
currently skipped.

b/205869701

Change-Id: Id363329a55b2f7e563c754e6ea6eb1647c880ffd