Set the reconciler pod's default security context
Autopilot applies the container runtime default seccomp profile to all
Pods in the cluster [1]. That makes the reconciler Deployment keeps
generating new replicasets on the Autopilot cluster (rapid channel)
because the Deployment's pod template is not the same as the
Replicaset's pod template. The Deployment's pod template doesn't have
the default `seccompProfile` security context.

This commit adds `setcommProfile` as the default security context so
that the reconciler won't fall into a terminating-creating loop.

[1]
https://cloud.google.com/kubernetes-engine/docs/concepts/autopilot-overview#container_isolation

b/212633129

Change-Id: I132a87ed6c7ae6e93a701f55868beb1f8d4c6870