Delete the ClusterRoleBinding when RootSync is deleted
ClusterRoleBinding configsync.gke.io:root-reconciler is a resource
created by the reconciler-manager. When RootSync is deleted, the
generaetd resource should be garbage collected.

We cannot rely on the OwnerReferences for the resource deletion because
ClusterRoleBinding is a cluster-scoped resource, while RootSync is a
namespace-scoped resources. Cluster-scoped dependents can only specify
cluster-scoped owners and cross-namespace owner references are
disallowed by design. Hence, this CL removes the OwnerReferences on the
ClusterRoleBinding object and adds a cleanup function to remove the
ClusterRoleBinding object.

b/215570929

Change-Id: Iaf724639b1771224beb004e9fc9a2f13810d7172