add e2e testcase to verify sync ordering
This change tests two cases in go/cs-sync-ordering

Case 16: object B depends on object A. Both A and B exist
in the Git repository, and don’t exist on the cluster.
Expected behavior: object A is applied and reconciled before object B is applied.

Case 17:object B depends on object A. Both A and B exist
in the Git repository, and don’t exist on the cluster.
Object A fails to reconcile (or times out)
Expected behavior: object A is applied. object B apply is skipped.
Current behavior: cli-utils does not verify whether B's
dependencies were applied and reconciled before applying.

Bug: b/217783459

Change-Id: I53592e073034edba20381ef1dd3882861e045689