Fix the CI test failures on gke clusters
All the GKE CI tests failed in the sync_ordering_test.go because it was
run after switch_mode_test.go, which switched to the mono-repo mode. The
RootSync object doesn't exist, so the setup logic creates a default one
with hierarchical source format.

This commit updates the resetRootRepoSpec function to create a RootSync
object if not found with the correct format.

It also resets the multi-repo mode so resources can be cleaned
gracefully.

b/218338842
b/218338959
b/218339029
b/218339005
b/218338934
b/218338977
b/218338975
b/218338984
b/218338952
b/218338931

Change-Id: Icba440b931440465c5687f4155ddeeb8a4f86a72