Allow declaring RootSync in root repos
The multi-sync feature supports declaring RootSync objects in the root
repository.

Currently, Config Sync forbids declaring the config-management-system
namespace, or resources within it:
https://cloud.google.com/anthos-config-management/docs/reference/errors#knv1034.
While, the RootSync object is a namespace-scoped resource in the
config-management-system namespace.

We have to loose the restriction on declaring resources in the
config-management-system namespace. We can still have the restriction
with the exception of RootSync.

b/218721425

Change-Id: I97073710ef7f0525193d3cac09c5179c098bbdcd