Update the scope of RootSync to NamespaceScope
Though RootSync appears to be a cluster-scoped resource, the CRD
defines the namespace, so it is actually namespace scoped.

`nomos_resources.txt` declares RootSync ClusterScoped. KNV1052 disallows
cluster-scoped resources to declare configs with `metadata.namespace`
field. That will block the support for declaring RootSync in root repos.

Hence, this CL updates RootSync to be NamespaceScoped.

b/219515227
Change-Id: I86579eab7c17b1153e4eb6d09d14ab29015583a1

Change-Id: I85ab4a76f1173112a983eb36852949dba7e7e9b9