add recursiveDiff function for e2e tests
This change consolidates the calls to the diff tool to a common
function. This is intended to provide a common abstraction for making
calls to diff.

Additionally, the function uses flags to ignore changes due to comments
and whitespace which are introduced when license headers are added to
the golden files.

Bug: b/216203702

Change-Id: Ifb8d7ac1b00a87c91d85c7e68f056e40d6a744b8