Attach RootSync|RepoSync CRs to the externally watched resources
The reconciler-manager generates resources for RootSync and RepoSync
CRs, and watches the managed resources. When they're updated,
it requeues the RooSync|RepoSync request to revert the changes.

Before the multi-sync feature, the reconciler-manager watches the
resources, extracts namespaces from the resources, maps the resource to
a RootSync|RepoSync request with the extracted namespace and a fixed
name using the mapper functions:
https://team.git.corp.google.com/gke-kubernetes-enterprise-control/prototype/+/refs/tags/v1.10.2-rc.1/pkg/reconcilermanager/controllers/mapper.go.

It doesn't work for multi-sync because the RootSync|RepoSync can have
arbitrary name, and would be hard to link to the corresponding CR from
the resource name. For example, if a Secret ssh-key is used by multiple
RepoSyncs in the same namespace. There is no way to attach to all
RepoSyncs from the Secret. Hence, we need to list the RootSync|RepoSync
CRs via API calls and use the .spec.git.secretRef.name field to check
the resource name.

It also fixes a secret mapping issue. The previous code uses a hard-coded
secret suffix (ssh-key). That is only used in the test, but not reflect
the real secretRef that is used in production.

b/219766771

Change-Id: Icb0f342f566b178a42220333aa8335b8abb5f7eb