addlicense to compiled files under examples
These files could not previously have license headers added because the
diff tool was fragile to diffs in comments. The tests have since been
changed to be robust to differences due to comments such as the license
header. The license header is being added to prepare for open sourcing
ConfigSync.

Bug: b/216203702

Change-Id: I8234ce1435277d515794547e89aa20592e8d01d5