Add implicit namespaces only when absent
The root parser adds implicit namespaces for unstructured format. If a
namespace-scoped resource is declared without a namespace in the root
repo, the parser will create an implicit namespace and add it to the
declared resources.

It works okay when there is only one Root reconciler, but now we are
supporting multiple RootSyncs. Adding the implicit namespace to the
declared resources list will cause management conflict among multiple
reconcilers.

The solution is to check the existence of the namespace. If the
namespace already exists with no manager or other manager except the
current reconciler, no-op to avoid conflict. Otherwise, creates an
implicit namespace. When other root reconcilers attempt to add the
same implicit namespace, there is one already exists and managed by
a different manager, other root reconcilers won't manage it again.

b/221496292

Change-Id: Ie1698c5d30314ddd56dda972c28d87db7b813ae9