Defer the parse-apply-watch loop when conflict is detected
If conflict is detected by the remediator, the reconciler retries the
parse-apply-watch loop in a second. It is highly likely the conflict
won't be resolved in a second. The next retry loop might succeed and
overwrite the management conflict error, which causes the sync status
bouncing between success and failure.

This commit adds delay before retrying the loop, so the conflict error
is more visible. The wait time is same as the polling frequency.

b/213343423

Change-Id: I538a7a1ef1fc11a539a663340c1b5532192c0d1c